Move pr-archive-creation-start-screen to `glam` folder
A `glam` subfolder for onboarding components was introduced in a5958c4.
This folder was intended to house the GLAM versions of components so
that the division between new and old components was clearer. Let's use
it.

PER-9753: Glam Onboarding Header
