+++ src/app/onboarding/components/glam/archive-creation-start-screen/archive-creation-start-screen.component.ts
