+++ src/app/onboarding/onboarding.module.ts
@@ -21,7 +21,7 @@ import { OnboardingComponent } from './components/onboarding/onboarding.componen
 import { WelcomeScreenComponent } from './components/welcome-screen/welcome-screen.component';
 import { CreateNewArchiveComponent } from './components/create-new-archive/create-new-archive.component';
 import { ArchiveTypeSelectComponent } from './components/archive-type-select/archive-type-select.component';
-import { ArchiveCreationStartScreenComponent } from './components/archive-creation-start-screen/archive-creation-start-screen.component';
+import { ArchiveCreationStartScreenComponent } from './components/glam/archive-creation-start-screen/archive-creation-start-screen.component';
 import { SelectArchiveTypeScreenComponent } from './components/select-archive-type-screen/select-archive-type-screen.component';
 import { NameArchiveScreenComponent } from './components/name-archive-screen/name-archive-screen.component';
 import { GlamArchiveTypeSelectComponent } from './components/glam/archive-type-select/archive-type-select.component';
